% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_functions.R
\name{aero_resist}
\alias{aero_resist}
\title{Aerodynamic resistance}
\usage{
aero_resist(uz, wind_z, z0, A, lake_z, rho_a = 1.2, ca = 0.001013)
}
\arguments{
\item{uz}{wind speed at wind_z height (m/s)}

\item{wind_z}{height at which \code{uz} is measured (m)}

\item{z0}{aerodynamic roughness of land cover at measurement site (m)}

\item{A}{surface area of the lake (km^2)}

\item{lake_z}{elevation of lake above mean sea level (m)}

\item{rho_a}{density of the air (kg/m^3), defaults to 1.20 kg/m^3 at 20 deg C}

\item{ca}{specific heat of the air (MJ/kg/K), defaults to 0.001013 MJ/kg/K}
}
\value{
\item{ra}{aerodynamic resistance (s/m)}
}
\description{
Calculates the aerodynamic resistance (s/m) over a lake as defined by Calder
and Neal (1984, pg. 93) and McJannet et al. (2008, Appendix B, Equation 10),
and as presented in McMahon et al. (2013) Equation S11.23.
}
\references{
McMahon, T. A., Peel, M. C., Lowe, L., Srikanthan, R., and
McVicar, T. R. (2013). Estimating actual, potential, reference crop and pan
evaporation using standard meteorological data: a pragmatic synthesis,
Hydrol. Earth Syst. Sci., 17, 1331–1363.
https://doi.org/10.5194/hess-17-1331-2013.

McJannet, D. L., Webster, I. T., Stenson, M. P., and Sherman,
B.S. (2008). Estimating open water evaporation for the Murray-Darling
Basin. A report to the Australian Government from the CSIRO Murray-Darling
Basin Sustainable Yields Project, CSIRO, Australia, 50 pp. Retrieved from
http://www.clw.csiro.au/publications/waterforahealthycountry/mdbsy/technical/U-OpenWaterEvaporation.pdf.
}
