% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSLS_find_common_dates.R
\name{CSLS_find_common_dates}
\alias{CSLS_find_common_dates}
\title{Given CSLS data, find common dates of record}
\usage{
CSLS_find_common_dates(weather, lake_levels, lst)
}
\arguments{
\item{weather}{a data frame with hourly weather data incl. air temperature
(atmp), relative humidity (RH), incoming solar radiation (Rs),
precipitation (P), and wind speed (wind).}

\item{lake_levels}{a data frame with daily water level measurements as
formatted in the lake_levels dataset,
subset to lake level records for the lake of interest.}

\item{lst}{a data frame with sub-monthly lake surface temperature
measurements as formatted in the lst_HOBO dataset, subset
for a single lake.}
}
\value{
:
\item{z}{elevation above mean sea level (m)}
\item{phi}{latitude of location (radians). Positive for northern
           hemisphere, negative for southern hemisphere}
\item{Lm}{longitude of location (degrees west of Greenwich)}
\item{Lz}{longitude of location's measurement timezone (degrees west of
            Greenwich). For example, Lz = 75, 90, 105 and 120째 for
            measurement times based on the Eastern, Central, Rocky Mountain
            and Pacific time zones (United States) and Lz = 0째 for Greenwich,
            330째 for Cairo (Egypt), and 255째 for Bangkok (Thailand).}
}
\description{
Finds date interval with records for lst, weather, and lake levels.
}
