% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSLS_format_lake.R
\name{CSLS_format_lake}
\alias{CSLS_format_lake}
\title{Given CSLS data, format "lake" input parameter}
\usage{
CSLS_format_lake(elev_area_vol, lake_levels, lst = NULL, wtmp0, use_lst)
}
\arguments{
\item{elev_area_vol}{a data frame with the lake, stage_m, surf_area_m2, and
volume_m3 as in the elev_area_voldataset, subset
for a single lake.}

\item{lake_levels}{a data frame with daily water level measurements as
formatted in the lake_levels dataset,
subset to lake level records for the lake of interest.}

\item{lst}{a data frame with sub-monthly lake surface temperature
measurements as formatted in the lst_HOBO dataset, subset
for a single lake.}

\item{wtmp0}{initial water temperature for first day in timeseries (degC)}

\item{use_lst}{logical defaults to TRUE to use available lake surface
temperature data.}
}
\value{
lake, a list with the following lake-specific parameters:
\item{A}{surface area of the lake (km^2)}
\item{depth_m}{depth of the lake (m). Can be a static value or vector
              corresponding with datetimes vector.}
\item{lst}{optional data frame with date (datetime) and ltmp (lake
           temperature, degC).}
\item{wtmp0}{required initial water temperature for first day in datetimes
             (degC)}
}
\description{
Extracts location information for the lake from the site dictionary and
reformats for input into lake evaporation functions.
}
