% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSLS_format_loc.R
\name{CSLS_format_loc}
\alias{CSLS_format_loc}
\title{Given CSLS data, format "loc" input parameter}
\usage{
CSLS_format_loc(dictionary, Lz = 90)
}
\arguments{
\item{dictionary}{a data frame with the obs_type ("LK"), elev_m, lat_deg,
and long_deg for the lake of interest.}

\item{Lz}{the longitude of the local timezone (degrees west of Greenwich,
ranges from 0 to 360 degrees). Defaults to 90 for Central Time
Zone, USA.}
}
\value{
*\emph{loc}, a list with the following location-specific parameters:
\item{z}{elevation above mean sea level (m)}
\item{phi}{latitude of location (radians). Positive for northern
hemisphere, negative for southern hemisphere}
\item{Lm}{longitude of location (degrees west of Greenwich)}
\item{Lz}{longitude of location's measurement timezone (degrees west of
Greenwich). For example, Lz = 75, 90, 105 and 120째 for
measurement times based on the Eastern, Central, Rocky Mountain
and Pacific time zones (United States) and Lz = 0째 for Greenwich,
330째 for Cairo (Egypt), and 255째 for Bangkok (Thailand).}
}
\description{
Extracts location information for the lake from the dictionary and reformats
for input into lake evaporation functions.
}
