% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSLS_format_weather.R
\name{CSLS_format_weather}
\alias{CSLS_format_weather}
\title{Given CSLS data, format "weather" input parameter}
\usage{
CSLS_format_weather(weather, wind_elev = 3, z0 = 0.02)
}
\arguments{
\item{weather}{a data frame with hourly weather data incl. air temperature
(atmp), relative humidity (RH), incoming solar radiation (Rs),
precipitation (P), and wind speed (wind).}

\item{wind_elev}{height at which wind is measured (m), default: 3}

\item{z0}{aerodynamic roughness of weather measurement site (m), default: 0.02}
}
\value{
weather a list with weather data that includes:
\item{dt}{string indicating the timestep of input weather series. Expects
"hourly", "daily", or "monthly".}
\item{datetimes}{datetimes of weather records \link{POSIXct}. If monthly timestep,
make sure date is the 15th of each month.}
\item{atmp}{If hourly timestep, vector of air temperature (degrees C)
corresponding with datetimes vector. If daily or monthly
timestep, list with two vectors, "min" and "max", with mean daily
min and max air temperature (degrees C) corresponding with
datetimes vector}
\item{RH}{If hourly timestep, vector of relative humidity (percent)
corresponding with datetimes vector. If daily or monthly timestep,
list with two vectors, "min" and "max", with mean daily min and max
relative humidity (percent) corresponding with datetimes vector.}
\item{Rs}{vector of incoming solar or shortwave radiation (MJ/m^2/hr if
hourly timestep, MG/m^2/day if daily or monthly), corresponding
with datetimes vector.}
\item{wind}{vector with mean wind speed (m/s), corresponding with datetimes
vector.}
\item{wind_elev}{atomic number, height at which wind is measured (m)}
\item{z0}{aerodynamic roughness of weather measurement site (m)}
}
\description{
Takes hourly weather data, summarizes as daily weather data, and re-formats
for input to lake evaporation functions
}
