% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_functions.R
\name{dew_tmp}
\alias{dew_tmp}
\title{Dew Point Temperature}
\usage{
dew_tmp(atmp, RH)
}
\arguments{
\item{atmp}{air temperature (degrees C). When dt is "daily" or larger,
argument should be a list with elements "min" and "max" for daily
min and max temperatures. When dt is "hourly", argument should be
a vector with hourly recorded temperatures.}

\item{RH}{relative humidity (percent). When dt is "daily" or larger, argument
should be a list with elements "min" and "max" for daily min and
max relative humidities. When dt is "hourly", argument should be a
vector with hourly recorded relative humidities.}
}
\value{
\item{dewtmp}{the dew point temperature (degrees C)}
}
\description{
Calculates the dew point temperature as a function of actual vapour pressure,
based on McJannet et al. (2008) Equation 26 as presented in McMahon et al.
(2013) Equation S2.3
}
\references{
McMahon, T. A., Peel, M. C., Lowe, L., Srikanthan, R., and
  McVicar, T. R.: Estimating actual, potential, reference crop and pan
  evaporation using standard meteorological data: a pragmatic synthesis,
  Hydrol. Earth Syst. Sci., 17, 1331–1363,
  https://doi.org/10.5194/hess-17-1331-2013, 2013.
}
