% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaporation.R
\name{evaporation}
\alias{evaporation}
\title{Evaporation (All Methods)}
\usage{
evaporation(method = "FAO", loc = NULL, weather, lake = NULL,
  albedo = list(ref_crop = 0.23, water = 0.08), no_condensation = TRUE,
  rho_a = 1.2, ca = 0.001013)
}
\arguments{
\item{method}{denotes which evaporation method to use ("FAO" or "McJannet").}

\item{loc}{a list with location information that includes:
\itemize{
\item z - elevation above mean sea level (m)
\item phi - latitude (radians). Positive for northern hemisphere, negative
            for southern hemisphere.
\item Lm - longitude of location (degrees west of Greenwich).
\item Lz - longitude of location's timezone (degrees west of Greenwich). For
           example, Lz = 75, 90, 105 and 120° for the Eastern, Central, Rocky
           Mountain and Pacific time zones (United States) and Lz = 0° for
           Greenwich, 330° for Cairo (Egypt), and 255° for Bangkok (Thailand)
}}

\item{weather}{a list with weather data that includes:
\itemize{
\item dt - string indicating the timestep of input weather series. Expects
           "hourly", "daily", or "monthly".
\item datetimes - datetimes of weather records [POSIXct]. If monthly
                  timestep, make sure date is the 15th of each month.
\item atmp - If hourly timestep, vector of air temperature (degrees C)
           corresponding with datetimes vector. If daily or monthly timestep,
           list with two vectors, "min" and "max", with mean daily min and
           max air temperature (degrees C) corresponding with datetimes
           vector
\item RH - If hourly timestep, vector of relative humidity (percent)
           corresponding with datetimes vector. If daily or monthly timestep,
           list with two vectors, "min" and "max", with mean daily min and
           max relative humidity (percent) corresponding with datetimes
           vector.
\item Rs - vector of incoming solar or shortwave radiation (MJ/m^2/timestep),
           corresponding with datetimes vector.
\item wind - vector with mean windspead (m/s), corresponding with datetimes
             vector.
\item wind_elev - atomic number, height at which wind is measured (m)
\item z0 - aerodynamic roughness of weather measurement site (m)
}}

\item{lake}{A list with lake data. Defaults to NULL, but for lake evaporation
            calculations should include:
\itemize{
  \item A - surface area of the lake (km^2).
  \item depth_m - depth of the lake (m).
  \item lst - data frame with date (datetime) and ltmp (lake temperature, degC)
  \item wtmp0 - initial water temperature for first day in datetimes.

}}

\item{albedo}{a list with albedos for different surfaces including (defaults):
\itemize{
  \item ref_crop - albedo of the hypothetical grass reference crop, 0.23
  \item water - albedo of water, 0.08.
}}

\item{no_condensation}{defaults to TRUE to force negative evapotranspiration
values (i.e., condensation) to zero}

\item{rho_a}{density of the air (kg/m^3), defaults to 1.20 kg/m^3 at 20 deg C}

\item{ca}{specific heat of the air (MJ/kg/K), defaults to 0.001013 MJ/kg/K}
}
\value{
\item{evap}{Evapo(transpi)ration (mm/hour or mm/day)}
}
\description{
Calculates evaporation using input data and specified calculation method.
Options include "FAO" Penman-Montieth reference evapotranspiration or
"McJannet" lake evaporation. Returns evaporation in mm/day, unless specify
"FAO" and submt hourly imput data, which returns evaporation in mm/hour.
}
\details{
"FAO" calculates FAO Penman-Monteith reference evapotranspiration (potential
evapotranspiration for a reference grass crop) based on Allen et al. (1998).

"McJannet" calculates daily evaporation from a lake based on the method in
McJannet et al. (2008) as presented in Equation S11.22 in McMahon et al.
(2013).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.

McJannet, D. L., Webster, I. T., Stenson, M. P., and Sherman,
  B.S. (2008). Estimating open water evaporation for the Murray-Darling
  Basin. A report to the Australian Government from the CSIRO Murray-Darling
  Basin Sustainable Yields Project, CSIRO, Australia, 50 pp. Retrieved from
  http://www.clw.csiro.au/publications/waterforahealthycountry/mdbsy/technical/U-OpenWaterEvaporation.pdf.

McMahon, T. A., Peel, M. C., Lowe, L., Srikanthan, R., and
  McVicar, T. R. (2013). Estimating actual, potential, reference crop and pan
  evaporation using standard meteorological data: a pragmatic synthesis,
  Hydrol. Earth Syst. Sci., 17, 1331–1363.
  https://doi.org/10.5194/hess-17-1331-2013.
}
