% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_functions.R
\name{FAO_eo}
\alias{FAO_eo}
\title{FAO Saturation Vapour Pressure}
\usage{
FAO_eo(tmp)
}
\arguments{
\item{tmp}{temperature of air or water (degrees C)}
}
\value{
eo - saturation vapour pressure at the temperature (kPa)
}
\description{
Calculates the saturation vapour pressure at a given temperature based on
equation 11 of Allen et al. (1998).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.
}
