% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_heat_functions.R
\name{FAO_G}
\alias{FAO_G}
\title{Soil Heat Flux - Hourly}
\usage{
FAO_G(dt, datetimes, phi, Lm, Lz, Rn, atmp)
}
\arguments{
\item{dt}{string indicating the timestep of input weather series. Expects
"hourly", "daily", or "monthly".}

\item{datetimes}{datetimes of weather records [POSIXct]. If monthly
timestep, make sure date is the 15th of each month.}

\item{phi}{latitude of the location (radians). Positive for northern
hemisphere, negative for southern hemisphere.}

\item{Lm}{longitude of the measurement site (degrees west of Greenwich)}

\item{Lz}{longitude of the center of the local time zone (degrees west of
Greenwich). For example, Lz = 75, 90, 105 and 120째 for the
Eastern, Central, Rocky Mountain and Pacific time zones (United
States) and Lz = 0째 for Greenwich, 330째 for Cairo (Egypt), and
255째 for Bangkok (Thailand).}

\item{Rn}{vector with net radiation (MJ/m^2/timestep)}

\item{atmp}{If hourly timestep, vector of air temperature (degrees C)
corresponding with datetimes vector. If daily or monthly timestep,
list with two vectors, "min" and "max", with mean daily min and
max air temperature (degrees C) corresponding with datetimes
vector}
}
\value{
\item{G}{soil heat flux (MJ/m^2/hr)}
}
\description{
Calculates soil heat flux for hourly, daily, or monthly calculations beneath a
dense cover of grass. Based on Equations 42-46 in Allen et al. (1998).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.
}
