% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation_functions.R
\name{FAO_inverse_dist}
\alias{FAO_inverse_dist}
\title{Inverse Relative Distance Earth-Sun}
\usage{
FAO_inverse_dist(J)
}
\arguments{
\item{J}{Julian day, i.e., the number of the day in the year between 1 (1
January) and 365 or 366 (31 December).}
}
\value{
\item{dr}{inverse relative distance Earth-Sun (unitless)}
}
\description{
Calculates the inverse relative distance Earth-Sun using Equation 23 from
Allen et al. (1998).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.
}
