% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_functions.R
\name{FAO_mean_ea}
\alias{FAO_mean_ea}
\title{FAO Mean Actual Vapour Pressure - Daily Timestep or Larger}
\usage{
FAO_mean_ea(dt, atmp, RH)
}
\arguments{
\item{dt}{a string indicating the timestep of calculation. If not "hourly",
assumes a daily or larger timestep}

\item{atmp}{air temperature (degrees C). When dt is "daily" or larger,
argument should be a list with elements "min" and "max" for daily
min and max temperatures. When dt is "hourly", argument should be
a vector with hourly recorded temperatures.}

\item{RH}{relative humidity (percent). When dt is "daily" or larger, argument
should be a list with elements "min" and "max" for daily min and
max relative humidities. When dt is "hourly", argument should be a
vector with hourly recorded relative humidities.}
}
\value{
\item{ea}{mean actual vapour pressure (kPa) during time period of
                  interest (hourly or daily timestep)}
}
\description{
Calculates the mean saturation vapour pressure for an hourly, daily, or
larger time period based on Equation 17 (for daily or larger timestep) or pg. of Allen et al. (1998).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.
}
