% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_functions.R
\name{FAO_mean_es}
\alias{FAO_mean_es}
\title{FAO Mean Saturation Vapour Pressure - Daily Timestep or Larger}
\usage{
FAO_mean_es(dt, atmp)
}
\arguments{
\item{dt}{a string indicating the timestep of calculation. If not "hourly",
assumes a daily or larger timestep}

\item{atmp}{air temperature (degrees C). When dt is "daily" or larger,
argument should be a list with elements "min" and "max" for daily
min and max temperatures. When dt is "hourly", argument should be
a vector with hourly recorded temperatures.}
}
\value{
\item{es}{mean saturation vapour pressure (kPa) during time period of
                  interest}
}
\description{
Calculates the mean saturation vapour pressure for a time period (daily
timestep or larger) using the mean daily minimum temperature and the mean
daily maximum temperature during that time period. Based on equation 12 of
Allen et al. (1998).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.
}
