% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation_functions.R
\name{FAO_Ra_night_adjusted}
\alias{FAO_Ra_night_adjusted}
\title{Nighttime Extraterrestrial Solar Radiation}
\usage{
FAO_Ra_night_adjusted(datetimes, phi, Lm, Lz, Gsc = 0.082)
}
\arguments{
\item{datetimes}{hourly timeseries to use in calculations [POSIXct].}

\item{phi}{latitude of the location (radians). Positive for northern
hemisphere, negative for southern hemisphere.}

\item{Lm}{longitude of the measurement site (degrees west of Greenwich)}

\item{Lz}{longitude of the center of the local time zone (degrees west of
Greenwich). For example, Lz = 75, 90, 105 and 120째 for the Eastern,
Central, Rocky Mountain and Pacific time zones (United States) and Lz = 0째
for Greenwich, 330째 for Cairo (Egypt), and 255째 for Bangkok (Thailand).}

\item{Gsc}{solar constant (MJ/m^2/min). Defaults to 0.0820 MJ/m^2/min.}
}
\value{
\item{Ra}{extraterrestrial radiation (MJ/m^2/hour) with nighttime Ra
                  values adjusted per Allen et al. (1998)}
}
\description{
Calculates the nighttime extraterrestrial radiation for a given location and
timestep. Based on recomendations on p. 75 in Allen et al. (1998) for hourly
ET caclculations.
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.
}
