% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_functions.R
\name{FAO_u2}
\alias{FAO_u2}
\title{FAO Wind Speed at 2m Height}
\usage{
FAO_u2(uz, z = 10)
}
\arguments{
\item{uz}{wind speed at z m above ground surface (m/s)}

\item{z}{height of measurement above ground surface (m). Default is 10m, a
common height for wind speed measurements in meteorology.}
}
\value{
\item{u2}{wind speed at 2 m above ground surface (m/s)}
}
\description{
Calculates the wind speed at 2m height given a wind speed at zm height based
on Equation 47 in Allen et al. (1998).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.
}
