% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_functions.R
\name{heat_flux}
\alias{heat_flux}
\title{Heat Flux}
\usage{
heat_flux(method, loc, lake, weather, Rn = NULL, rho_w = 997.9,
  cw = 0.00419)
}
\arguments{
\item{rho_w}{density of water (kg/m^3), defaults to 997.9 kg/m^3 at 20 deg C}

\item{cw}{specific heat of water (MJ/kg/K), defaults to 0.00419}
}
\value{
\item{G}{heat flux (MJ/m^2/hr)}
}
\description{
Calculates soil heat flux for hourly, daily, or monthly calculations beneath a
dense cover of grass. Based on Equations 42-46 in Allen et al. (1998).
}
\details{
Calculates the change in heat storage based on water temperature following
McJannet et al. (2008) Equation 31 as presented in McMahon et al. S11.33
(2013).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.

McMahon, T. A., Peel, M. C., Lowe, L., Srikanthan, R., and
  McVicar, T. R.: Estimating actual, potential, reference crop and pan
  evaporation using standard meteorological data: a pragmatic synthesis,
  Hydrol. Earth Syst. Sci., 17, 1331–1363,
  https://doi.org/10.5194/hess-17-1331-2013, 2013.
}
