% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_functions.R
\name{lake_Gw}
\alias{lake_Gw}
\title{Change in Heat Storage}
\usage{
lake_Gw(loc, lake, weather, rho_w = 997.9, cw = 0.00419)
}
\arguments{
\item{loc}{a list with location information that includes:
\itemize{
  \item z - elevation above mean sea level (m)
  \item phi - latitude (radians). Positive for northern hemisphere, negative
            for southern hemisphere.
  \item Lm - longitude of location (degrees west of Greenwich). Only needed if
           input data is hourly.
  \item Lz - longitude of location's timezone (degrees west of Greenwich).
             For example, Lz = 75, 90, 105 and 120° for the Eastern, Central,
             Rocky Mountain and Pacific time zones (United States) and Lz =
             0° for Greenwich, 330° for Cairo (Egypt), and 255° for Bangkok
             (Thailand). Only needed if input is hourly.
}}

\item{lake}{a list with lake data that includes:
\itemize{
  \item A - surface area of the lake (km^2).
  \item depth_m - depth of the lake (m).
  \item lst - data frame with date (datetime) and ltmp (lake temperature, degC)
}}

\item{weather}{a list with weather data that includes:
\itemize{
  \item dt - string indicating the timestep of input weather series. Expects
             "hourly", "daily", or "monthly".
  \item datetimes - datetimes of weather records [POSIXct]. If monthly
                    timestep, make sure date is the 15th of each month.
  \item atmp - If hourly timestep, vector of air temperature (degrees C)
               corresponding with datetimes vector. If daily or monthly
               timestep, list with two vectors, "min" and "max", with mean
               daily min and max air temperature (degrees C) corresponding
               with datetimes vector.
  \item RH - If hourly timestep, vector of relative humidity (percent)
             corresponding with datetimes vector. If daily or monthly
             timestep, list with two vectors, "min" and "max", with mean
             daily min and max relative humidity (percent) corresponding with
             datetimes vector.
  \item Rs - vector of incoming solar or shortwave radiation (MJ/m^2/timestep),
             corresponding with datetimes vector.
  \item wind - vector with mean windspead (m/s), corresponding with datetimes
               vector.
  \item wind_elev - atomic number, height at which wind is measured (m)
  \item z0 - aerodynamic roughness of measurement site (m)
  \item wtmp0 - initial water temperature for first day in datetimes.
}}

\item{rho_w}{density of water (kg/m^3), defaults to 997.9 kg/m^3 at 20 deg C}

\item{cw}{specific heat of water (MJ/kg/K), defaults to 0.00419}
}
\value{
\describe{
\item{Gw}{change in heat storage (MJ/m^2/day)}
}
}
\description{
Calculates the change in heat storage based on water temperature following
McJannet et al. (2008) Equation 31 as presented in McMahon et al. S11.33
(2013).
}
\references{
McMahon, T. A., Peel, M. C., Lowe, L., Srikanthan, R., and
  McVicar, T. R.: Estimating actual, potential, reference crop and pan
  evaporation using standard meteorological data: a pragmatic synthesis,
  Hydrol. Earth Syst. Sci., 17, 1331–1363,
  https://doi.org/10.5194/hess-17-1331-2013, 2013.
}
