% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation_functions.R
\name{lake_Rol}
\alias{lake_Rol}
\title{Outgoing Longwave Radiation at Water Temperature}
\usage{
lake_Rol(wtmp, SBc = 4.903e-09)
}
\arguments{
\item{wtmp}{daily water temperature (degrees C).}

\item{SBc}{Stefan-Boltzman constant (MJ/m^2/day). Defaults to 4.903e-9
MJ/m^2/day.}
}
\value{
{Rol}{outgoing longwaver radiation (MJ/m^2/day) at the temperature of
              the water}
}
\description{
Calculates the outgoing longwave radiation for a lake as a function of wet
bulb temperature based on McMahon et al. (2013) Equation S11.27.
}
\references{
McMahon, T. A., Peel, M. C., Lowe, L., Srikanthan, R., and
  McVicar, T. R.: Estimating actual, potential, reference crop and pan
  evaporation using standard meteorological data: a pragmatic synthesis,
  Hydrol. Earth Syst. Sci., 17, 1331–1363,
  https://doi.org/10.5194/hess-17-1331-2013, 2013.
}
