% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_functions.R
\name{latent_heat_vapor}
\alias{latent_heat_vapor}
\title{Latent Heat of Vaporization}
\usage{
latent_heat_vapor(atmp)
}
\arguments{
\item{atmp}{air temperature (degrees C)}
}
\value{
\item{lambda}{the latent heat of vaporization (MJ/kg)}
}
\description{
Calculates the latent heat of vaporization as a function of air temperature.
}
\references{
\url{https://cran.r-project.org/web/packages/bigleaf/bigleaf.pdf}
}
