% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_radiation_functions.R
\name{McJannet_Cf}
\alias{McJannet_Cf}
\title{Cloudiness Factor}
\usage{
McJannet_Cf(Rs, Rso)
}
\arguments{
\item{Rs}{incoming solar radiation (MJ/m^2/day)}

\item{Rso}{clear sky radiation (MJ/m^2/day).}
}
\value{
{Cf}{the fraction of cloud cover (-)}
}
\description{
Estimates the fraction of cloud cover based on McJannet et al. (2008b,
Equations 14 and 15), as presented by McMahon et al. (2013), Equations S3.17
and S3.18.
}
\references{
McMahon, T. A., Peel, M. C., Lowe, L., Srikanthan, R., and
  McVicar, T. R.: Estimating actual, potential, reference crop and pan
  evaporation using standard meteorological data: a pragmatic synthesis,
  Hydrol. Earth Syst. Sci., 17, 1331–1363,
  https://doi.org/10.5194/hess-17-1331-2013, 2013.
}
