% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_heat_functions.R
\name{McJannet_eqtmp}
\alias{McJannet_eqtmp}
\title{Equilibrium Temperature}
\usage{
McJannet_eqtmp(loc, lake, weather, SBc = 4.903e-09)
}
\arguments{
\item{loc}{a list with location information that includes:
\itemize{
\item z - elevation above mean sea level (m)
}}

\item{lake}{a list with lake data that includes:
\itemize{
\item A - surface area of the lake (km^2).
\item depth_m - depth of the lake (m).
}}

\item{weather}{a list with weather data that includes:
\itemize{
\item dt - string indicating the timestep of input weather series. Expects
           "hourly", "daily", or "monthly".
\item atmp - If hourly timestep, vector of air temperature (degrees C)
           corresponding with datetimes vector. If daily or monthly timestep,
           list with two vectors, "min" and "max", with mean daily min and
           max air temperature (degrees C) corresponding with datetimes
           vector
\item RH - If hourly timestep, vector of relative humidity (percent)
           corresponding with datetimes vector. If daily or monthly timestep,
           list with two vectors, "min" and "max", with mean daily min and
           max relative humidity (percent) corresponding with datetimes
           vector.
\item wind - vector with mean windspead (m/s), corresponding with datetimes
             vector.
\item wind_elev - atomic number, height at which wind is measured (m)
\item z0 - aerodynamic roughness of measurement site (m)
}}

\item{SBc}{Stefan-Boltzman constant (MJ/m^2/day). Defaults to 4.903e-9
MJ/m^2/day.}
}
\value{
{eqtmp}{equilibrium temperature (degrees C)}
}
\description{
Calculates the equilibrium temperature based on de Bruin (1982) Equation 3,
as presented in McMahon et al. (2013) S11.30.
}
\references{
McMahon, T. A., Peel, M. C., Lowe, L., Srikanthan, R., and
  McVicar, T. R.: Estimating actual, potential, reference crop and pan
  evaporation using standard meteorological data: a pragmatic synthesis,
  Hydrol. Earth Syst. Sci., 17, 1331–1363,
  https://doi.org/10.5194/hess-17-1331-2013, 2013.
}
