% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_heat_functions.R
\name{McJannet_time_const}
\alias{McJannet_time_const}
\title{Time Constant}
\usage{
McJannet_time_const(loc, lake, weather, rho_w = 997.9, cw = 0.00419,
  SBc = 4.903e-09)
}
\arguments{
\item{loc}{a list with location information that includes:
\itemize{
\item z - elevation above mean sea level (m)
}}

\item{lake}{a list with lake data that includes:
\itemize{
\item A - surface area of the lake (km^2).
\item depth_m - depth of the lake (m).
}}

\item{weather}{a list with weather data that includes:
\itemize{
\item dt - string indicating the timestep of input weather series. Expects
           "hourly", "daily", or "monthly".
\item atmp - If hourly timestep, vector of air temperature (degrees C)
           corresponding with datetimes vector. If daily or monthly timestep,
           list with two vectors, "min" and "max", with mean daily min and
           max air temperature (degrees C) corresponding with datetimes
           vector
\item RH - If hourly timestep, vector of relative humidity (percent)
           corresponding with datetimes vector. If daily or monthly timestep,
           list with two vectors, "min" and "max", with mean daily min and
           max relative humidity (percent) corresponding with datetimes
           vector.
\item wind - vector with mean windspead (m/s), corresponding with datetimes
             vector.
\item wind_elev - atomic number, height at which wind is measured (m)
\item z0 - aerodynamic roughness of measurement site (m)
}}

\item{rho_w}{density of water (kg/m^3), defaults to 997.9 kg/m^3 at 20 deg C}

\item{cw}{specific heat of water (MJ/kg/K), defaults to 0.00419}

\item{SBc}{Stefan-Boltzman constant (MJ/m^2/day). Defaults to 4.903e-9
MJ/m^2/day.}
}
\value{
{Ctime}{time constant (day)}
}
\description{
Calculates the time constant (day) based on McJannet et al. (2008) Equation 5
and de Bruin (1982) Equation 4, as presented in McMahon et al. (2013) S11.29.
}
\references{
McMahon, T. A., Peel, M. C., Lowe, L., Srikanthan, R., and
  McVicar, T. R.: Estimating actual, potential, reference crop and pan
  evaporation using standard meteorological data: a pragmatic synthesis,
  Hydrol. Earth Syst. Sci., 17, 1331–1363,
  https://doi.org/10.5194/hess-17-1331-2013, 2013.
}
