% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_heat_functions.R
\name{McJannet_wtmp}
\alias{McJannet_wtmp}
\title{Water Temperature}
\usage{
McJannet_wtmp(loc, lake, weather)
}
\arguments{
\item{loc}{a list with location information that includes:
\itemize{
\item z - elevation above mean sea level (m)
\item phi - latitude (radians). Positive for northern hemisphere, negative
            for southern hemisphere.
\item Lm - longitude of location (degrees west of Greenwich).
\item Lz - longitude of location's timezone (degrees west of Greenwich). For
           example, Lz = 75, 90, 105 and 120° for the Eastern, Central, Rocky
           Mountain and Pacific time zones (United States) and Lz = 0° for
           Greenwich, 330° for Cairo (Egypt), and 255° for Bangkok (Thailand)
}}

\item{lake}{a list with lake data that includes:
\itemize{
\item A - surface area of the lake (km^2).
\item depth_m - depth of the lake (m).
}}

\item{weather}{a list with weather data that includes:
\itemize{
\item dt - string indicating the timestep of input weather series. Expects
           "hourly", "daily", or "monthly".
\item datetimes - datetimes of weather records [POSIXct]. If monthly
                  timestep, make sure date is the 15th of each month.
\item atmp - If hourly timestep, vector of air temperature (degrees C)
           corresponding with datetimes vector. If daily or monthly timestep,
           list with two vectors, "min" and "max", with mean daily min and
           max air temperature (degrees C) corresponding with datetimes
           vector
\item RH - If hourly timestep, vector of relative humidity (percent)
           corresponding with datetimes vector. If daily or monthly timestep,
           list with two vectors, "min" and "max", with mean daily min and
           max relative humidity (percent) corresponding with datetimes
           vector.
\item Rs - vector of incoming solar or shortwave radiation (MJ/m^2/timestep),
           corresponding with datetimes vector.
\item wind - vector with mean windspead (m/s), corresponding with datetimes
             vector.
\item wind_elev - atomic number, height at which wind is measured (m)
\item z0 - aerodynamic roughness of measurement site (m)
\item wtmp0 - initial water temperature for first day in datetimes.
}}
}
\value{
{wtmp}{water temperature (degrees C)}
}
\description{
Calculates the water temperature based on the water temperature of the
previous day following McJannet et al. (2008) Equation 23 and de Bruin (1982)
Equation 10 as presented in McMahon et al. S11.28 (2013).
}
\references{
McMahon, T. A., Peel, M. C., Lowe, L., Srikanthan, R., and
  McVicar, T. R.: Estimating actual, potential, reference crop and pan
  evaporation using standard meteorological data: a pragmatic synthesis,
  Hydrol. Earth Syst. Sci., 17, 1331–1363,
  https://doi.org/10.5194/hess-17-1331-2013, 2013.
}
