% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_functions.R
\name{psychrometric_constant}
\alias{psychrometric_constant}
\title{Psychrometric Constant}
\usage{
psychrometric_constant(z, atmp = NULL, lambda = 2.45, cp = 0.001013,
  epsilon = 0.622)
}
\arguments{
\item{z}{elevation above sea level (m)}

\item{atmp}{air temperature (degrees C). Optional value used to calculate the
latent heat of vaporization. Defaults to NULL to use constant
value of lambda.}

\item{lambda}{latent heat of vaporization (MJ/kg). Default is 2.45.}

\item{cp}{specific heat at constant pressure (MJ/kg/degC). Default is
1.1013e-3.}

\item{epsilon}{ratio molecular weight of water vapour/dry air (-). Default is
0.622}
}
\value{
\item{gamma}{the psychrometric constant for the given elevation
                     (kPa/degrees C)}
}
\description{
Calculates the psychrometric constant for a given elevation based on Equation
7 and 8 in Allen et al. (1998).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.
}
