% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation_functions.R
\name{R_n}
\alias{R_n}
\title{Radiation, net}
\usage{
R_n(type = "FAO", loc, lake, weather, albedo)
}
\arguments{
\item{type}{type of net radiation function to use. Defaults to "FAO" for FAO
Penman-Monteith reference evapotranspiration. Other options
include "McJannet" for net radiation at lake temperature and
"wet_bulb" for net radiation at wet-bulb temperature.}

\item{loc}{a list with location information that includes:
\itemize{
\item "z" - elevation above mean sea level (m)
\item"phi" - latitude of location (radians). Positive for northern
             hemisphere, negative for southern hemisphere.
\item"Lm" - longitude of location (degrees west of Greenwich).
\item"Lz" - longitude of location's measurement timezone (degrees west of
            Greenwich). For example, Lz = 75, 90, 105 and 120° for
            measurement times based on the Eastern, Central, Rocky Mountain
            and Pacific time zones (United States) and Lz = 0° for Greenwich,
            330° for Cairo (Egypt), and 255° for Bangkok (Thailand).
}}

\item{lake}{A list with lake data. Defaults to NULL, but for lake evaporation
            calculations, should include:
\itemize{
  \item "A" - surface area of the lake (km^2).
  \item "depth_m" - depth of the lake (m). Can be a static value or vector
                    corresponding with datetimes vector.
  \item "lst" - optional data frame with date (datetime) and ltmp (lake
                temperature, degC).
  \item "wtmp0" - required initial water temperature for first day in
                  datetimes (degC).

}}

\item{weather}{a list with weather data that includes:
\itemize{
\item "dt" - string indicating the timestep of input weather series. Expects
             "hourly", "daily", or "monthly".
\item "datetimes" - datetimes of weather records [POSIXct]. If monthly
                    timestep, make sure date is the 15th of each month.
\item "atmp" - If hourly timestep, vector of air temperature (degrees C)
               corresponding with datetimes vector. If daily or monthly
               timestep, list with two vectors, "min" and "max", with mean
               daily min and max air temperature (degrees C) corresponding
               with datetimes vector
\item "RH" - If hourly timestep, vector of relative humidity (percent)
             corresponding with datetimes vector. If daily or monthly
             timestep, list with two vectors, "min" and "max", with mean
             daily min and max relative humidity (percent) corresponding with
             datetimes vector.
\item "Rs" - vector of incoming solar or shortwave radiation (MJ/m^2/hr if
             hourly timestep, MG/m^2/day if daily or monthly), corresponding
             with datetimes vector.
\item "wind" - vector with mean wind speed (m/s), corresponding with
               datetimes vector.
\item "wind_elev" - atomic number, height at which wind is measured (m)
\item "z0" - aerodynamic roughness of weather measurement site (m)
}}

\item{albedo}{a list with albedos for different surfaces, defaults to:
\itemize{
  \item "ref_crop" - albedo of the hypothetical grass reference crop, 0.23
  \item "water" - albedo of water, 0.08.
}}
}
\value{
\item{Rn}{net radiation (MJ/m^2/timestep)}
}
\description{
Calculates the net solar or shortwave radiation using the "FAO" Penman
Monteith approach for reference evapotranspiration (Allen et al., 1998,
Equation 38), the "McJannet" approach for lake temperature (McMahon et al.,
2013, Equation S11.25), or the McJannet approach for "wet_bulb" temperature
(McMahon et al., 2013, Equation S11.31).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.

McMahon, T. A., Peel, M. C., Lowe, L., Srikanthan, R., and
  McVicar, T. R. (2013). Estimating actual, potential, reference crop and pan
  evaporation using standard meteorological data: a pragmatic synthesis,
  Hydrol. Earth Syst. Sci., 17, 1331–1363.
  https://doi.org/10.5194/hess-17-1331-2013.
}
