% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation_functions.R
\name{R_nl}
\alias{R_nl}
\title{Radiation, net longwave}
\usage{
R_nl(dt, Rs, Rso, ea, atmp, SBc = 4.903e-09)
}
\arguments{
\item{dt}{string indicating the timestep of input weather series. Expects
"hourly", "daily", or "monthly".}

\item{Rs}{incoming solar radiation (MJ/m^2/timestep), vector or atomic number}

\item{Rso}{clear-sky solar radiation (MJ/m^2/timestep), vector or atomic
number.}

\item{ea}{actual vapour pressure (kPa), vector or atomic number.}

\item{atmp}{air temperature (degrees C). When dt is "daily" or larger,
argument should be a list with elements "min" and "max" for daily
min and max temperatures. When dt is "hourly", argument should be
a vector or atomic number with hourly recorded temperatures.}

\item{SBc}{Stefan-Boltzman constant (MJ/m^2/day). Defaults to 4.903e-9
MJ/m^2/day.}
}
\value{
\item{Rnl}{Net outgoing longwave radiation (MJ/m^2/hour or
                   MJ/m^2/day, depending on dt)}
}
\description{
Calculates the net outgoing longwave radiation for a given location based on
incoming solar radiation, incoming clear-sky radiation, and air temperature
records. Based on Equation 39 and p.74-75 in Allen et al. (1998).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.
}
