% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation_functions.R
\name{R_ns}
\alias{R_ns}
\title{Net Solar or Shortwave Radiation}
\usage{
R_ns(Rs, albedo = 0.23)
}
\arguments{
\item{Rs}{incoming solar radiation (MJ/m^2/timestep)}

\item{albedo}{albedo or canopy reflection coefficient (-). Defaults to 0.23
for the hypothetical grass reference crop.}
}
\value{
\item{Rns}{Net solar or shortwave radiation (MJ/m^2/timestep)}
}
\description{
Calculates the net solar or shortwave radiation for a given location based on
incoming solar radiation timeseries. Based on Equation 38 in Allen et al.
(1998).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.
}
