% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation_functions.R
\name{R_ol}
\alias{R_ol}
\title{Outgoing Longwave Radiation at Wet Bulb Temperature}
\usage{
R_ol(type = "McJannet", tmp1, tmp2 = NULL, SBc = 4.903e-09)
}
\arguments{
\item{type}{type of outgoing longwave radiation function to use. Defaults to
"McJannet" for outgoing longwave radiation at lake temperature.
Set to "wet_bulb" for outgoing longwave radiation at wet-bulb
temperature.}

\item{tmp1}{daily lake water temperature (if "McJannet") or wet bulb
temperature (if "wet_bulb") (degrees C).}

\item{tmp2}{defaults to NULL. If "wet_bulb", set to air temperature (degrees
C). When dt is "daily" or larger, argument should be a list with
elements "min" and "max" for daily min and max temperatures. When
dt is "hourly", argument should be a vector with hourly recorded
temperatures.}

\item{SBc}{Stefan-Boltzman constant (MJ/m^2/day). Defaults to 4.903e-9
MJ/m^2/day.}
}
\value{
{Rol}{outgoing longwave radiation (MJ/m^2/day)}
}
\description{
Calculates the outgoing longwave radiation for a lake as a function of lake
temperature ("McJannet"; McMahon et al., 2013, Equation S11.27) or as a
function of wet bulb temperature  ("wet_bulb"; McMahon et al., 2013, Equation
S11.32).
}
\references{
McMahon, T. A., Peel, M. C., Lowe, L., Srikanthan, R., and
  McVicar, T. R.: Estimating actual, potential, reference crop and pan
  evaporation using standard meteorological data: a pragmatic synthesis,
  Hydrol. Earth Syst. Sci., 17, 1331–1363,
  https://doi.org/10.5194/hess-17-1331-2013, 2013.
}
