% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation_functions.R
\name{R_so}
\alias{R_so}
\title{Radiation, clear-sky solar}
\usage{
R_so(Ra, z)
}
\arguments{
\item{Ra}{extraterrestrial solar radiation (MJ/m^2/day)}

\item{z}{station elevation above sea level (m).}
}
\value{
\item{Rso}{clear-sky solar radiation (MJ/m^2/day)}
}
\description{
Calculates the clear-sky solar radiation based on Equation 37 in Allen et al.
(1998).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.
}
