% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_functions.R
\name{sat_vapor_density}
\alias{sat_vapor_density}
\title{Saturation vapor density}
\usage{
sat_vapor_density(atmp)
}
\arguments{
\item{atmp}{air temperature (degrees C). When dt is "daily" or larger,
argument should be a list with elements "min" and "max" for daily
min and max temperatures. When dt is "hourly", argument should be
a vector with hourly recorded temperatures.}
}
\value{
\item{svd}{saturated vapor density (g/m^3)}
}
\description{
Calculates the saturation vapor density from the ideal gas law based on
Equation 2.6 in Harwell (2012).
}
\references{
Harwell, G.R., 2012, Estimation of evaporation from open water—A
  review of selected studies, summary of U.S. Army Corps of Engineers data
  collection and methods, and evaluation of two methods for estimation of
  evaporation from five reservoirs in Texas: U.S. Geological Survey
  Scientific Investigations Report 2012–5202, 96 p.
}
