% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_functions.R
\name{time_const}
\alias{time_const}
\title{Time constant}
\usage{
time_const(loc, lake, weather, rho_w = 997.9, cw = 0.00419, SBc = 4.903e-09)
}
\arguments{
\item{loc}{a list with location information that includes:
\itemize{
\item "z" - elevation above mean sea level (m)
\item"phi" - latitude of location (radians). Positive for northern
             hemisphere, negative for southern hemisphere.
\item"Lm" - longitude of location (degrees west of Greenwich).
\item"Lz" - longitude of location's measurement timezone (degrees west of
            Greenwich). For example, Lz = 75, 90, 105 and 120° for
            measurement times based on the Eastern, Central, Rocky Mountain
            and Pacific time zones (United States) and Lz = 0° for Greenwich,
            330° for Cairo (Egypt), and 255° for Bangkok (Thailand).
}}

\item{lake}{A list with lake data. Defaults to NULL, but for McJannet lake
            evaporation calculations, should include:
\itemize{
  \item "A" - surface area of the lake (km^2).
  \item "depth_m" - depth of the lake (m). Can be a static value or vector
                    corresponding with datetimes vector.
  \item "lst" - optional data frame with date (datetime) and ltmp (lake
                temperature, degC).
  \item "wtmp0" - required initial water temperature for first day in
                  datetimes (degC).

}}

\item{weather}{a list with weather data that includes:
\itemize{
\item "dt" - string indicating the timestep of input weather series. Expects
             "hourly", "daily", or "monthly".
\item "datetimes" - datetimes of weather records [POSIXct]. If monthly
                    timestep, make sure date is the 15th of each month.
\item "atmp" - If hourly timestep, vector of air temperature (degrees C)
               corresponding with datetimes vector. If daily or monthly
               timestep, list with two vectors, "min" and "max", with mean
               daily min and max air temperature (degrees C) corresponding
               with datetimes vector
\item "RH" - If hourly timestep, vector of relative humidity (percent)
             corresponding with datetimes vector. If daily or monthly
             timestep, list with two vectors, "min" and "max", with mean
             daily min and max relative humidity (percent) corresponding with
             datetimes vector.
\item "Rs" - vector of incoming solar or shortwave radiation (MJ/m^2/hr if
             hourly timestep, MG/m^2/day if daily or monthly), corresponding
             with datetimes vector.
\item "wind" - vector with mean wind speed (m/s), corresponding with
               datetimes vector.
\item "wind_elev" - atomic number, height at which wind is measured (m)
\item "z0" - aerodynamic roughness of weather measurement site (m)
}}

\item{rho_w}{density of water (kg/m^3), defaults to 997.9 kg/m^3 at 20 deg C}

\item{cw}{specific heat of water (MJ/kg/K), defaults to 0.00419}

\item{SBc}{Stefan-Boltzman constant (MJ/m^2/day). Defaults to 4.903e-9
MJ/m^2/day.}
}
\value{
\item{Ctime}{time constant (day)}
}
\description{
Calculates the time constant (day) based on McJannet et al. (2008) Equation 5
and de Bruin (1982) Equation 4, as presented in McMahon et al. (2013) S11.29.
}
\references{
McMahon, T. A., Peel, M. C., Lowe, L., Srikanthan, R., and
  McVicar, T. R.: Estimating actual, potential, reference crop and pan
  evaporation using standard meteorological data: a pragmatic synthesis,
  Hydrol. Earth Syst. Sci., 17, 1331–1363,
  https://doi.org/10.5194/hess-17-1331-2013, 2013.
}
