% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_functions.R
\name{tmp_water}
\alias{tmp_water}
\title{Temperature, lake water}
\usage{
tmp_water(loc, lake, weather, albedo)
}
\arguments{
\item{loc}{a list with location information that includes:
\itemize{
\item "z" - elevation above mean sea level (m)
\item"phi" - latitude of location (radians). Positive for northern
             hemisphere, negative for southern hemisphere.
\item"Lm" - longitude of location (degrees west of Greenwich).
\item"Lz" - longitude of location's measurement timezone (degrees west of
            Greenwich). For example, Lz = 75, 90, 105 and 120° for
            measurement times based on the Eastern, Central, Rocky Mountain
            and Pacific time zones (United States) and Lz = 0° for Greenwich,
            330° for Cairo (Egypt), and 255° for Bangkok (Thailand).
}}

\item{lake}{A list with lake data. Defaults to NULL, but for McJannet lake
            evaporation calculations, should include:
\itemize{
  \item "A" - surface area of the lake (km^2).
  \item "depth_m" - depth of the lake (m). Can be a static value or vector
                    corresponding with datetimes vector.
  \item "lst" - optional data frame with date (datetime) and ltmp (lake
                temperature, degC).
  \item "wtmp0" - required initial water temperature for first day in
                  datetimes (degC).

}}

\item{weather}{a list with weather data that includes:
\itemize{
\item "dt" - string indicating the timestep of input weather series. Expects
             "hourly", "daily", or "monthly".
\item "datetimes" - datetimes of weather records [POSIXct]. If monthly
                    timestep, make sure date is the 15th of each month.
\item "atmp" - If hourly timestep, vector of air temperature (degrees C)
               corresponding with datetimes vector. If daily or monthly
               timestep, list with two vectors, "min" and "max", with mean
               daily min and max air temperature (degrees C) corresponding
               with datetimes vector
\item "RH" - If hourly timestep, vector of relative humidity (percent)
             corresponding with datetimes vector. If daily or monthly
             timestep, list with two vectors, "min" and "max", with mean
             daily min and max relative humidity (percent) corresponding with
             datetimes vector.
\item "Rs" - vector of incoming solar or shortwave radiation (MJ/m^2/hr if
             hourly timestep, MG/m^2/day if daily or monthly), corresponding
             with datetimes vector.
\item "wind" - vector with mean wind speed (m/s), corresponding with
               datetimes vector.
\item "wind_elev" - atomic number, height at which wind is measured (m)
\item "z0" - aerodynamic roughness of weather measurement site (m)
}}

\item{albedo}{a list with albedos for different surfaces, defaults to:
\itemize{
  \item "ref_crop" - albedo of the hypothetical grass reference crop, 0.23
  \item "water" - albedo of water, 0.08.
}}
}
\value{
\item{wtmp}{water temperature (degrees C)}
}
\description{
Calculates the water temperature based on the water temperature of the
previous day following McJannet et al. (2008) Equation 23 and de Bruin (1982)
Equation 10 as presented in McMahon et al. S11.28 (2013). If measured lake
surface temperatures are provided (in \code{lake$lst}), will force water
temperature to those values on all days with measurements and estimate any
missing values using the McJannet and de Bruin appoach.
}
\references{
McMahon, T. A., Peel, M. C., Lowe, L., Srikanthan, R., and
  McVicar, T. R. (2013). Estimating actual, potential, reference crop and pan
  evaporation using standard meteorological data: a pragmatic synthesis,
  Hydrol. Earth Syst. Sci., 17, 1331–1363.
  https://doi.org/10.5194/hess-17-1331-2013.

McJannet, D. L., Webster, I. T., Stenson, M. P., and Sherman,
  B.S. (2008). Estimating open water evaporation for the Murray-Darling
  Basin. A report to the Australian Government from the CSIRO Murray-Darling
  Basin Sustainable Yields Project, CSIRO, Australia, 50 pp. Retrieved from
  http://www.clw.csiro.au/publications/waterforahealthycountry/mdbsy/technical/U-OpenWaterEvaporation.pdf.
}
