% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_functions.R
\name{uz_to_u10}
\alias{uz_to_u10}
\title{Wind Speed at 10m Height}
\usage{
uz_to_u10(uz, z = 2, z0 = 0.02)
}
\arguments{
\item{uz}{wind speed at z m above ground surface (m/s)}

\item{z}{height of measurement above ground surface (m). Default is 2m, a
common height for wind speed measurements in agronomy}

\item{z0}{roughness length for the measurement surface (m). Defaults to 0.02
for a short grass.}
}
\value{
\item{u10}{wind speed at 10 m above ground surface (m/s)}
}
\description{
Calculates the wind speed at 10m height given a wind speed at zm height based
on Equation S4.4 in McMahon et al. (2013).
}
\references{
McMahon, T. A., Peel, M. C., Lowe, L., Srikanthan, R., and
  McVicar, T. R.: Estimating actual, potential, reference crop and pan
  evaporation using standard meteorological data: a pragmatic synthesis,
  Hydrol. Earth Syst. Sci., 17, 1331–1363,
  https://doi.org/10.5194/hess-17-1331-2013, 2013.
}
