% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_functions.R
\name{vp_act_mean}
\alias{vp_act_mean}
\title{Mean Actual Vapour Pressure}
\usage{
vp_act_mean(atmp, RH)
}
\arguments{
\item{atmp}{air temperature (degrees C). When dt is "daily" or larger,
argument should be a list with elements "min" and "max" for daily
min and max temperatures. When dt is "hourly", argument should be
a vector with hourly recorded temperatures.}

\item{RH}{relative humidity (percent). When dt is "daily" or larger, argument
should be a list with elements "min" and "max" for daily min and
max relative humidities. When dt is "hourly", argument should be a
vector with hourly recorded relative humidities.}
}
\value{
\item{ea}{mean actual vapour pressure (kPa) during time period of
                  interest (hourly or daily timestep)}
}
\description{
Calculates the mean actual vapour pressure for an hourly, daily, or larger
time period based on Equation 17 (for daily or larger timestep) on pg. 38 of
Allen et al. (1998).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.
}
