% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_functions.R
\name{vp_deficit}
\alias{vp_deficit}
\title{Vapour Pressure Deficit}
\usage{
vp_deficit(atmp, RH)
}
\arguments{
\item{atmp}{air temperature (degrees C). When dt is "daily" or larger,
argument should be a list with elements "min" and "max" for daily
min and max temperatures. When dt is "hourly", argument should be
a vector with hourly recorded temperatures.}

\item{RH}{relative humidity (percent). When dt is "daily" or larger, argument
should be a list with elements "min" and "max" for daily min and
max relative humidities. When dt is "hourly", argument should be a
vector with hourly recorded relative humidities.}
}
\value{
\item{vpd}{vapour pressure deficit (kPa) for the given time period}
}
\description{
Calculates the vapour pressure deficit for a given time period (daily timestep
or larger) using the mean daily minimum temperature, mean daily maximum
temperature, mean daily minimum relative humidity, and mean daily maxiumum
humidity during that time period. See p.39 in Allen et al. (1998).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.
}
