% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_functions.R
\name{vp_sat_curve_slope}
\alias{vp_sat_curve_slope}
\title{Slope of saturation vapour pressure curve}
\usage{
vp_sat_curve_slope(atmp)
}
\arguments{
\item{atmp}{air temperature (degrees C). When dt is "daily" or larger,
argument should be a list with elements "min" and "max" for daily
min and max temperatures. When dt is "hourly", argument should be
a vector with hourly recorded temperatures.}
}
\value{
\item{Delta}{slope of the saturation vapour pressure curve at the
given air temperature}
}
\description{
Calculates the slope of the saturation vapour ressure curve (i.e., the slope
of the relationship between saturation vapour pressure and temperature) based
on Equation 13 in Allen et al. (1998).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
evapotranspiration: Guidelines for computing crop water requirements. Rome:
FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.
}
