% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_functions.R
\name{vp_sat_mean}
\alias{vp_sat_mean}
\title{Mean Saturation Vapour Pressure}
\usage{
vp_sat_mean(atmp)
}
\arguments{
\item{atmp}{air temperature (degrees C). When dt is "daily" or larger,
argument should be a list with elements "min" and "max" for daily
min and max temperatures. When dt is "hourly", argument should be
a vector with hourly recorded temperatures.}
}
\value{
\item{es}{mean saturation vapour pressure (kPa) during time period of
                  interest}
}
\description{
Calculates the mean saturation vapour pressure for an hourly, daily, or
larger time period based on Equation 12 (for daily or larger timestep) on pg.
36 of Allen et al. (1998).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.
}
