% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation_functions.R
\name{wet_bulb_Rol}
\alias{wet_bulb_Rol}
\title{Outgoing Longwave Radiation at Wet Bulb Temperature}
\usage{
wet_bulb_Rol(atmp, wbtmp, SBc = 4.903e-09)
}
\arguments{
\item{atmp}{air temperature (degrees C). When dt is "daily" or larger,
argument should be a list with elements "min" and "max" for daily
min and max temperatures. When dt is "hourly", argument should be
a vector with hourly recorded temperatures.}

\item{wbtmp}{daily wet bulb temperature (degrees C).}

\item{SBc}{Stefan-Boltzman constant (MJ/m^2/day). Defaults to 4.903e-9
MJ/m^2/day.}
}
\value{
{Rolwb}{outgoing longwaver radiation (MJ/m^2/day) at wet-bulb
                temperature}
}
\description{
Calculates the outgoing longwave radiation for a lake as a function of wet
bulb temperature based on McMahon et al. (2013) Equation S11.32.
}
\references{
McMahon, T. A., Peel, M. C., Lowe, L., Srikanthan, R., and
  McVicar, T. R.: Estimating actual, potential, reference crop and pan
  evaporation using standard meteorological data: a pragmatic synthesis,
  Hydrol. Earth Syst. Sci., 17, 1331–1363,
  https://doi.org/10.5194/hess-17-1331-2013, 2013.
}
