% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_functions.R
\name{wind_fcn}
\alias{wind_fcn}
\title{Wind Speed at 10m Height}
\usage{
wind_fcn(u10, A)
}
\arguments{
\item{u10}{wind speed at 10 m above ground surface (m/s)}

\item{A}{area of the water body (km^2)}
}
\value{
\item{u_fcn}{wind function from McJannet}
}
\description{
Calculates the wind function given the wind speed at 10m and the surface area
of the lake, based on McJannet et al. (2008) Appendix B, Equation 10, as
presented in Equation S11.24 in McMahon et al. (2013).
}
\references{
McMahon, T. A., Peel, M. C., Lowe, L., Srikanthan, R., and
  McVicar, T. R.: Estimating actual, potential, reference crop and pan
  evaporation using standard meteorological data: a pragmatic synthesis,
  Hydrol. Earth Syst. Sci., 17, 1331–1363,
  https://doi.org/10.5194/hess-17-1331-2013, 2013.
}
