% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSLS_daily_met.R
\name{CSLS_daily_met}
\alias{CSLS_daily_met}
\title{Given CSLS lake and method, calculate daily evaporation}
\usage{
CSLS_daily_met(
  method = "McJannet",
  use_lst = TRUE,
  lakename,
  Lz = 90,
  wind_elev = 3,
  z0 = 0.02,
  no_condensation = FALSE
)
}
\arguments{
\item{method}{denotes which evaporation method to use ("FAO", "McJannet", or
"Hamon").}

\item{use_lst}{logical defaults to TRUE to use available lake surface
temperature data.}

\item{lakename}{name of lake currently evaluating}

\item{Lz}{longitude of location's measurement timezone (degrees west of
Greenwich). For example, Lz = 75, 90, 105 and 120° for measurement
times based on the Eastern, Central, Rocky Mountain and Pacific
time zones (United States).}

\item{wind_elev}{height at which wind is measured (m), default: 3}

\item{z0}{aerodynamic roughness of weather measurement site (m), default: 0.2}

\item{no_condensation}{defaults to FALSE. If TRUE, forces negative
evapotranspiration values (i.e., condensation) to zero}
}
\value{
\strong{weather}, a data frame with daily weather information including:
\item{date}{day of each weather observation \link{POSIXct}}
\item{atmp_min}{minimum air temperature for the day (deg C)}
\item{atmp_max}{maximum air temperature for the day (deg C)}
\item{RH_min}{minimum relative humidity for the day (percent)}
\item{RH_max}{maximum relative humidity for the day (percent)}
\item{P}{total precipitation for the day (mm)}
\item{E}{total lake evaporation for the day (mm)}
}
\description{
This function calculates daily evaporation for a CSLS lake given the desired
method and lake name. While this function can use any of the methods
availaible in \code{\link{evaporation}}, it is currently hardwired to load
input data from the CSLS field campaign (as available in \pkg{CSLSdata}) and
analysis is restricted to dates with available weather, lake surface
temperature, and lake level data (via \code{\link{CSLS_find_common_dates}})
regardless of the requirements of the chosen method.
}
\examples{
# Note that "lakename" is not really used for "Hamon" method
# However, this function is currently hardwired to require it.
lake      <- "Pleasant"
daily_met <- CSLS_daily_met("McJannet", lakename = lake)
daily_met <- CSLS_daily_met("McJannet", use_lst = FALSE, lakename = lake)
daily_met <- CSLS_daily_met("Hamon", lakename = lake)

}
