% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSLS_find_common_dates.R
\name{CSLS_find_common_dates}
\alias{CSLS_find_common_dates}
\title{Given CSLS data, find common dates of record}
\usage{
CSLS_find_common_dates(weather, lake_levels, lst)
}
\arguments{
\item{weather}{a data frame with hourly weather data incl. air temperature
(atmp), relative humidity (RH), incoming solar radiation (Rs),
precipitation (P), and wind speed (wind).}

\item{lake_levels}{a data frame with daily water level measurements as
formatted in the lake_levels dataset,
subset to lake level records for the lake of interest.}

\item{lst}{a data frame with sub-monthly lake surface temperature
measurements as formatted in the lst_HOBO dataset, subset
for a single lake.}
}
\value{
a list with the following items:
\item{weather}{same as input weather data frame, but filtered to only include
data during the overlap time period}
\item{lst}{same as input lst data frame, but filtered to only include data
during the overlap time period and summarized at a daily time
step}
\item{lake_levels}{same as input lake_levels data frame, but filtered to only
include data during the overlap time period and with
missing (NA) lake levels interpolated}
\item{wtmp0}{initial lake surface temperatuer (degC) from the day before the
weather, lst, and lake_levels timeseries begin}
}
\description{
Finds date interval with records for lst, weather, and lake levels. Checks
for the latest "start" date in each of these timeseries, then checks for the
earliest "end" date in each timeseries, then restricts all input data to be
within the overlapping interval. Lastly, interpolates the \code{lake_levels}
dataset over \code{NA} values to ensure a continuous timeseries of lake area and
lake depth for evaporation calculations.
}
\details{
Note that this function may cause undesired behavior when using
\code{\link{CSLS_daily_met}} with the Unmodified Hammon method or the
McJannet method with \code{use_lst = FALSE}. Neither uses the \code{lst}
dataset for calculations, yet other input data is restricted based on the
availability of the \code{lst} dataset. Unmodified Hamon also does not use
\code{lake_levels}, but is restricted by the avilability of that dataset as
well.
}
