% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation_functions.R
\name{hour_angle_sunset}
\alias{hour_angle_sunset}
\title{Sunset Hour Angle}
\usage{
hour_angle_sunset(phi, delta)
}
\arguments{
\item{phi}{the latitude of the location (radians). Positive for northern
hemisphere, negative for southern hemisphere.}

\item{delta}{the solar declination based on the Julian day (radians)}
}
\value{
\item{omega_s}{sunset hour angle (radians)}
}
\description{
Calculates the sunset hour angle (radians) for a given location and day of
the year using Equation 25 from Allen et al. (1998).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop
  evapotranspiration: Guidelines for computing crop water requirements. Rome:
  FAO. Retrieved from http://www.fao.org/docrep/X0490E/x0490e00.htm.
}
